refactor: Dropzone/LayoutDropzoneをFileInput.tsに統合
