+++ src/components/FileInput.ts
@@ -1,6 +1,28 @@
+import { readFileAsText } from "../lib/csv";
 import { parseLayout, buildLayoutMeta, type Layout, type LayoutMeta } from "../lib/layout";
 
-export function initLayoutDropzone(
+export function initCsvInput(
+  onLoad: (csvText: string, fileName: string) => void,
+  onError: (msg: string) => void
+): void {
+  const fileInput = document.getElementById("file-input") as HTMLInputElement;
+
+  fileInput.addEventListener("change", (e) => {
+    const f = (e.target as HTMLInputElement).files?.[0];
+    if (f) handleFile(f);
+  });
+
+  async function handleFile(file: File) {
+    try {
+      const csvText = await readFileAsText(file);
+      onLoad(csvText, file.name);
+    } catch (err) {
+      onError((err as Error).message);
+    }
+  }
+}
+
+export function initLayoutInput(
   onLoad: (layout: Layout, meta: LayoutMeta, fileName: string) => void,
   onError: (msg: string) => void
 ): void {
